refactor: Created `NewBreedViewModel` class to separate from original `BreedListViewModel`
